requirements-gralaf.txt
causalnex==0.11.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
dill==0.3.5.1
fonttools==4.34.4
idna==3.3
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
multiprocess==0.70.13
networkx==2.8.5
numpy==1.22.4
packaging==21.3
pandas==1.4.3
pathos==0.2.9
patsy==0.5.2
pgmpy==0.1.19
Pillow==9.2.0
pox==0.3.1
ppft==1.7.6.5
pygraphviz==1.10
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
statsmodels==0.13.2
threadpoolctl==3.1.0
torch==1.12.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
wrapt==1.12.1
matplotlib==3.6.1
